backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib
argon2-cffi
python-jose
pydantic
PyJWT